requirements.txt
ctransformers
transformers
langchain
langchain-community
torch
pypdf
sentence-transformers
gpt4all
faiss-cpu
openai
PyPDF2
streamlit
bardapi
pillow
matplotlib
pandas
plotly
requests
keras
scikit-learn
tabulate
librosa
scipy